// ui/src/components/Typography/types.ts
export type TypographyVariant =
    | 'h1'
    | 'h2'
    | 'h3'
    | 'h4'
    | 'h5'
    | 'b1'
    | 'b2'
    | 'b3'
    | 'b4'
    | 'b5'
    | 'sm'
    | 'md'
    | 'lg'
    | 'xl'
    | 'title'
    | 'button'
    | 'logo'

export type TypographyTextStyle =
    | 'text-heading-1'
    | 'text-heading-2'
    | 'text-heading-3'
    | 'text-heading-4'
    | 'text-heading-5'
    | 'text-body-1'
    | 'text-body-2'
    | 'text-body-3'
    | 'text-body-4'
    | 'text-body-5'
    | 'text-body-xl'
    | 'text-body-lg'
    | 'text-body-sm'
    | 'text-body-md'
    | 'text-body-title'
    | 'text-button'
    | 'text-logo'

export type TextColor =
    | 'gradient'
    | 'default'
    | 'primary'
    | 'secondary'
    | 'tertiary'
    | 'quaternary'
    | 'success'
    | 'error'
    | 'dark'
    | 'gray'
    | 'none'